Completed 9, 10, 11

+++ src/p10/summationofprimes/main.java
@@ -5,8 +5,7 @@ package p10.summationofprimes;
  */
 public class main {
     public static void main(String[] args) {
-        int sum = 2;
-
+        long sum = 2;
 
         for (int i = 3; i < 2000000; i+=2) {
             if (p10.prime10001st.main.checkPrime(i)) sum += i;
